libraries/Microsoft.Bot.Builder.Runtime/Providers/Telemetry/ApplicationInsightsTelemetryProvider.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.

using System;
using AdaptiveExpressions.Properties;
using Microsoft.ApplicationInsights.Extensibility;
using Microsoft.Bot.Builder;
using Microsoft.Bot.Builder.ApplicationInsights;
using Microsoft.Bot.Builder.Integration.ApplicationInsights.Core;
using Microsoft.Bot.Builder.Runtime.Extensions;
using Microsoft.Extensions.Configuration;
using Microsoft.Extensions.DependencyInjection;
using Newtonsoft.Json;

namespace Microsoft.Bot.Builder.Runtime.Providers.Telemetry
{
    [JsonObject]
    public class ApplicationInsightsTelemetryProvider : ITelemetryProvider
    {
        [JsonProperty("$kind")]
        public const string Kind = "Microsoft.ApplicationInsightsTelemetryProvider";

        [JsonProperty("instrumentationKey")]
        public StringExpression InstrumentationKey { get; set; }

        public void ConfigureServices(IServiceCollection services, IConfiguration configuration)
        {
            if (services == null)
            {
                throw new ArgumentNullException(nameof(services));
            }

            if (configuration == null)
            {
                throw new ArgumentNullException(nameof(configuration));
            }

            services.AddApplicationInsightsTelemetry(this.InstrumentationKey?.GetConfigurationValue(configuration));
            services.AddSingleton<ITelemetryInitializer, OperationCorrelationTelemetryInitializer>();
            services.AddSingleton<ITelemetryInitializer, TelemetryBotIdInitializer>();
            services.AddSingleton<IBotTelemetryClient, BotTelemetryClient>();
        }
    }
}
